requirements.txt
flask
python-dotenv
gunicorn
requests
pyjwt
deep-translator
